@@ -1,12 +1,14 @@
+from __future__ import annotations
+
 from collections import deque
 from time import monotonic
-from typing import Deque, Optional
+from typing import Deque
 
 from spiel.constants import TARGET_RPS
 
 
 class RPSCounter:
-    def __init__(self, render_history_length: Optional[int] = None) -> None:
+    def __init__(self, render_history_length: int | None = None) -> None:
         if render_history_length is None:
             render_history_length = 3 * TARGET_RPS
 
