@@ -1,14 +1,17 @@
+from __future__ import annotations
+
+from collections.abc import Iterable, Iterator
 from itertools import zip_longest
-from typing import Any, Iterable, Iterator, Optional, TypeVar
+from typing import TypeVar
 
 T = TypeVar("T")
 
 
-def filter_join(separator: str, items: Iterable[Optional[Any]]) -> str:
+def filter_join(separator: str, items: Iterable[object | None]) -> str:
     return separator.join(map(str, filter(None, items)))
 
 
-def drop_nones(*items: Optional[T]) -> Iterator[T]:
+def drop_nones(*items: T | None) -> Iterator[T]:
     yield from (item for item in items if item is not None)
 
 
@@ -20,6 +23,6 @@ def clamp(value: int, lower: int, upper: int) -> int:
     return max(min(value, upper), lower)
 
 
-def chunks(iterable: Iterable[T], n: int, fill_value: Optional[T] = None) -> Iterable[Iterable[T]]:
+def chunks(iterable: Iterable[T], n: int, fill_value: T | None = None) -> Iterable[Iterable[T]]:
     args = [iter(iterable)] * n
     return zip_longest(*args, fillvalue=fill_value)
