@@ -1,4 +1,6 @@
-from typing import List, Optional, TypeVar
+from __future__ import annotations
+
+from typing import TypeVar
 
 import pytest
 
@@ -16,5 +18,5 @@ T = TypeVar("T")
         ("", 2, None, []),
     ],
 )
-def test_chunks(items: List[T], n: int, fill: Optional[T], expected: List[List[T]]) -> None:
+def test_chunks(items: List[T], n: int, fill: T | None, expected: list[list[T]]) -> None:
     assert [list(chunk) for chunk in chunks(items, n, fill_value=fill)] == expected
