@@ -1,6 +1,7 @@
+from collections.abc import Mapping
 from dataclasses import asdict, dataclass, fields
 from pathlib import Path
-from typing import Any, Mapping
+from typing import Any
 
 import tomli
 import tomli_w
